Added authorization for to users

+++ routers/posts.py
@@ -6,6 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
 import models
+from auth import CurrentUser
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
 
@@ -23,19 +24,12 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     return posts
 
 @router.post("",response_model=PostResponse,status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession,Depends(get_db)]):
-    result= await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user=result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="user Not Found"
-        )
+async def create_post(post: PostCreate, current_user:CurrentUser,db: Annotated[AsyncSession,Depends(get_db)]):
     
     new_post=models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -51,34 +45,36 @@ async def get_post_byid(post_id: int,db: Annotated[AsyncSession, Depends(get_db)
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Id not found")
 
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate ,db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id: int, current_user:CurrentUser,post_data: PostCreate ,db: Annotated[AsyncSession, Depends(get_db)]):
     result=await db.execute(select(models.Post).where(models.Post.id == post_id))
     post=result.scalars().first()
     if not post: 
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Post not found")
 
-    if post_data.user_id != post.user_id:
-        result= await db.execute(
-            select(models.User).where(models.User.id == post_data.user_id),
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update post"
         )
-        user=result.scalars().first()
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="User Not Found")
     
     post.title=post_data.title
     post.content=post_data.content
-    post.user_id=post_data.user_id
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
     return post
 
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate ,db: Annotated[AsyncSession, Depends(get_db)]):
-    result= await db.execute(select(models.Post).where(models.Post.id == post_id))
+async def update_post_partial(post_id: int, current_user:CurrentUser, post_data: PostUpdate ,db: Annotated[AsyncSession, Depends(get_db)]):
+    result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post=result.scalars().first()
     if not post: 
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Post not found")
     
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update post"
+        )
     # if post_data.title :
     #     post.title = post_data.title
     # if post_data.content :
@@ -92,7 +88,7 @@ async def update_post_partial(post_id: int, post_data: PostUpdate ,db: Annotated
     return post
 
 @router.delete("/{post_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int,current_user:CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result= await db.execute(select(models.Post).where(models.Post.id == post_id))
     post=result.scalars().first()
     if not post:
@@ -100,5 +96,11 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post Not found"
         )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to Delete post"
+        )
+    
     await db.delete(post)
     await db.commit()
